+++ internal/usecase/goal_usecase.go
@@ -6,8 +6,8 @@ import "context"
 type GoalUsecase interface {
 	List(ctx context.Context, userID int) ([]*GoalOutput, error)
 	Create(ctx context.Context, userID int, input *GoalInput) (*GoalOutput, error)
-	Update(ctx context.Context, userID int, id int64, input *GoalInput) (*GoalOutput, error)
-	Delete(ctx context.Context, userID int, id int64) error
+	Update(ctx context.Context, userID int, id uint32, input *GoalInput) (*GoalOutput, error)
+	Delete(ctx context.Context, userID int, id uint32) error
 }
 
 // GoalInput は目標の作成・更新入力です。
@@ -21,7 +21,7 @@ type GoalInput struct {
 
 // GoalOutput は目標API向けの出力です。
 type GoalOutput struct {
-	ID                int64
+	ID                uint32
 	Title             string
 	AchievementType   string
 	AchievementParams map[string]any
